backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
openai>=1.58.0
networkx==3.4.2
python-dotenv==1.0.1
